+++ app/components/UI/Predict/components/PredictGameChart/PredictGameChart.tsx
@@ -9,10 +9,8 @@ import { PredictGameStatus, PredictPriceHistoryInterval } from '../../types';
 import { usePredictPriceHistory } from '../../hooks/usePredictPriceHistory';
 import { useLiveMarketPrices } from '../../hooks/useLiveMarketPrices';
 import { usePredictGame } from '../../hooks/usePredictGame';
-import {
-  getPrimaryMoneylineOutcomes,
-  isDrawCapableLeague,
-} from '../../constants/sports';
+import { isDrawCapableLeague } from '../../constants/sports';
+import { getPrimaryMoneylineOutcomes } from '../../utils/sports';
 import { getLeagueTeamOrder } from '../../utils/gameParser';
 import { useTheme } from '../../../../../util/theme';
 import PredictGameChartContent from './PredictGameChartContent';
