+++ app/components/UI/Predict/constants/sports.ts
@@ -1,10 +1,4 @@
-import type {
-  PredictMarketGame,
-  PredictOutcome,
-  PredictOutcomeToken,
-  PredictSportTeam,
-  PredictSportsLeague,
-} from '../types';
+import type { PredictMarketGame, PredictSportsLeague } from '../types';
 
 /**
  * Leagues with live game data support.
@@ -170,377 +164,4 @@ export const shouldShowRegTimeTag = ({
   }
 
   return !nonRegTimeSportsMarketTypes.includes(sportsMarketType.toLowerCase());
-};
-
-interface NegRiskSportsMarket {
-  negRisk?: boolean;
-  sportsMarketType?: string;
-  groupItemTitle?: string;
-}
-
-interface TeamMatchedOutcome {
-  groupItemTitle?: string;
-  tokens: { title?: string }[];
-}
-
-export interface SportCardOutcome<TToken extends SportCardToken> {
-  id: string;
-  sportsMarketType?: string;
-  groupItemThreshold?: number;
-  tokens: TToken[];
-}
-
-export interface SportCardToken {
-  id: string;
-  title?: string;
-}
-
-export interface SportCardBetSide<
-  TOutcome extends SportCardOutcome<TToken>,
-  TToken extends SportCardToken,
-> {
-  outcome: TOutcome;
-  token: TToken;
-}
-
-export interface ResolvedSportCardButtons<
-  TOutcome extends SportCardOutcome<TToken>,
-  TToken extends SportCardToken,
-> {
-  home?: SportCardBetSide<TOutcome, TToken>;
-  draw?: SportCardBetSide<TOutcome, TToken>;
-  away?: SportCardBetSide<TOutcome, TToken>;
-  isTeamToAdvance: boolean;
-  remainingOptions: number;
-}
-
-interface SportsTeamLogoMarket extends NegRiskSportsMarket {
-  sportsMarketType?: string;
-}
-
-export const hasNegRiskMoneylineGroupItem = <T extends NegRiskSportsMarket>(
-  market: T,
-): market is T & { groupItemTitle: string } =>
-  Boolean(
-    market.negRisk &&
-      isMoneylineLikeMarketType(market.sportsMarketType) &&
-      market.groupItemTitle,
-  );
-
-const normalizeTeamLabel = (value?: string): string | undefined =>
-  value?.trim().toLowerCase();
-
-export const sportTeamMatchesLabel = (
-  label: string | undefined,
-  team: PredictSportTeam,
-): boolean => {
-  const normalizedLabel = normalizeTeamLabel(label);
-  if (!normalizedLabel) {
-    return false;
-  }
-
-  return [team.name, team.alias, team.abbreviation]
-    .map(normalizeTeamLabel)
-    .some((teamLabel) => teamLabel === normalizedLabel);
-};
-
-export const outcomeMatchesTeam = <T extends TeamMatchedOutcome>(
-  outcome: T,
-  team: PredictSportTeam,
-): boolean =>
-  sportTeamMatchesLabel(outcome.groupItemTitle, team) ||
-  sportTeamMatchesLabel(outcome.tokens[0]?.title, team);
-
-export const getTeamOutcome = <T extends TeamMatchedOutcome & { id: string }>(
-  outcomes: T[],
-  team: PredictSportTeam,
-  fallbackIndex: number,
-  excludedOutcome?: T,
-): T | undefined =>
-  outcomes.find(
-    (outcome) =>
-      outcome.id !== excludedOutcome?.id && outcomeMatchesTeam(outcome, team),
-  ) ??
-  outcomes.find((outcome) => outcome.id !== excludedOutcome?.id) ??
-  outcomes[fallbackIndex];
-
-export const getMatchingSportTeam = (
-  groupItemTitle: string,
-  game: PredictMarketGame,
-): PredictSportTeam | undefined =>
-  [game.homeTeam, game.awayTeam].find((team) =>
-    sportTeamMatchesLabel(groupItemTitle, team),
-  );
-
-const isGenericTeamLabel = (label: string): boolean => {
-  const normalizedLabel = normalizeTeamLabel(label);
-  return (
-    normalizedLabel === 'team to advance' ||
-    normalizedLabel?.startsWith('draw') === true
-  );
-};
-
-export const resolveNegRiskMoneylineShortTitles = (
-  market: NegRiskSportsMarket,
-  game: PredictMarketGame,
-): { yesShort?: string; noShort?: string } => {
-  if (!hasNegRiskMoneylineGroupItem(market)) {
-    return {};
-  }
-
-  if (market.groupItemTitle.toLowerCase().startsWith('draw')) {
-    return { yesShort: 'Draw' };
-  }
-
-  const yesTeam = getMatchingSportTeam(market.groupItemTitle, game);
-  if (!yesTeam) return {};
-
-  const isHome = yesTeam.id === game.homeTeam.id;
-  const noAbbr = isHome
-    ? game.awayTeam.abbreviation
-    : game.homeTeam.abbreviation;
-
-  return { yesShort: yesTeam.abbreviation, noShort: noAbbr };
-};
-
-export const getNegRiskMoneylineTeamLogo = (
-  market: NegRiskSportsMarket,
-  game?: PredictMarketGame,
-): string | undefined => {
-  if (!game || !hasNegRiskMoneylineGroupItem(market)) {
-    return undefined;
-  }
-
-  if (market.groupItemTitle.toLowerCase().startsWith('draw')) {
-    return undefined;
-  }
-
-  return getMatchingSportTeam(market.groupItemTitle, game)?.logo;
-};
-
-const hasSportsMarketTeamGroupItem = <T extends SportsTeamLogoMarket>(
-  market: T,
-): market is T & { groupItemTitle: string } =>
-  Boolean(
-    market.groupItemTitle &&
-      (hasNegRiskMoneylineGroupItem(market) ||
-        isTeamToAdvanceMarketType(market.sportsMarketType)),
-  );
-
-export const getSportsMarketTeamLogo = (
-  market: SportsTeamLogoMarket,
-  game?: PredictMarketGame,
-): string | undefined => {
-  if (!game || !hasSportsMarketTeamGroupItem(market)) {
-    return undefined;
-  }
-
-  if (isGenericTeamLabel(market.groupItemTitle)) {
-    return undefined;
-  }
-
-  return getMatchingSportTeam(market.groupItemTitle, game)?.logo;
-};
-
-export const getTeamToAdvanceTokenLogo = (
-  tokenTitle: string | undefined,
-  game?: PredictMarketGame,
-): string | undefined => {
-  if (
-    !game?.homeTeam ||
-    !game?.awayTeam ||
-    !tokenTitle ||
-    isGenericTeamLabel(tokenTitle)
-  ) {
-    return undefined;
-  }
-
-  return getMatchingSportTeam(tokenTitle, game)?.logo;
-};
-
-export const getTokenImage = ({
-  sportsMarketType,
-  tokenTitle,
-  game,
-}: {
-  sportsMarketType?: string;
-  tokenTitle?: string;
-  game?: PredictMarketGame;
-}): string | undefined => {
-  if (isTeamToAdvanceMarketType(sportsMarketType)) {
-    return getTeamToAdvanceTokenLogo(tokenTitle, game);
-  }
-
-  return undefined;
-};
-
-export const getBuyOutcomeImage = ({
-  outcome,
-  outcomeToken,
-  game,
-}: {
-  outcome: PredictOutcome;
-  outcomeToken?: PredictOutcomeToken;
-  game?: PredictMarketGame;
-}): string | undefined => {
-  if (
-    game?.league === WORLD_CUP_LEAGUE &&
-    isTeamToAdvanceMarketType(outcome.sportsMarketType)
-  ) {
-    return (
-      outcomeToken?.image ??
-      getTokenImage({
-        sportsMarketType: outcome.sportsMarketType,
-        tokenTitle: outcomeToken?.title,
-        game,
-      }) ??
-      outcome.image
-    );
-  }
-
-  return outcome.image;
-};
-
-export const getPrimaryMoneylineOutcomes = <
-  T extends { sportsMarketType?: string },
->(
-  outcomes: T[],
-): T[] => {
-  const moneylineOutcomes = outcomes.filter(
-    (outcome) => outcome.sportsMarketType?.toLowerCase() === 'moneyline',
-  );
-
-  return moneylineOutcomes.length > 0 ? moneylineOutcomes : outcomes;
-};
-
-export const getPrimarySportsCardOutcomes = <
-  T extends { sportsMarketType?: string },
->(
-  outcomes: T[],
-  league?: PredictSportsLeague,
-): T[] => {
-  if (league === WORLD_CUP_LEAGUE) {
-    const advanceOutcomes = outcomes.filter((outcome) =>
-      isTeamToAdvanceMarketType(outcome.sportsMarketType),
-    );
-
-    if (advanceOutcomes.length > 0) {
-      return advanceOutcomes;
-    }
-  }
-
-  return getPrimaryMoneylineOutcomes(outcomes);
-};
-
-const getTeamToken = <TToken extends SportCardToken>(
-  tokens: TToken[],
-  team: PredictSportTeam,
-  excludedTokenIds: string[] = [],
-): TToken | undefined =>
-  tokens.find((token) => sportTeamMatchesLabel(token.title, team)) ??
-  tokens.find((token) => !excludedTokenIds.includes(token.id)) ??
-  tokens[0];
-
-const getDrawToken = <TToken extends SportCardToken>(
-  tokens: TToken[],
-): TToken | undefined =>
-  tokens.find((token) => token.title?.toLowerCase() === 'draw');
-
-export const resolveSportCardButtons = <
-  TOutcome extends SportCardOutcome<TToken>,
-  TToken extends SportCardToken,
->({
-  outcomes,
-  game,
-  showDraw,
-}: {
-  outcomes: TOutcome[];
-  game: PredictMarketGame;
-  showDraw: boolean;
-}): ResolvedSportCardButtons<TOutcome, TToken> => {
-  const primaryOutcomes = getPrimarySportsCardOutcomes(outcomes, game.league);
-  const firstOutcome = primaryOutcomes[0];
-  const isTeamToAdvance =
-    game.league === WORLD_CUP_LEAGUE &&
-    isTeamToAdvanceMarketType(firstOutcome?.sportsMarketType);
-
-  const fallbackResult = {
-    isTeamToAdvance,
-    remainingOptions: isTeamToAdvance ? 0 : Math.max(0, outcomes.length - 1),
-  };
-
-  if (!firstOutcome) {
-    return fallbackResult;
-  }
-
-  const sortedDrawOutcomes =
-    showDraw && !isTeamToAdvance && primaryOutcomes.length >= 3
-      ? [...primaryOutcomes].sort(
-          (a, b) => (a.groupItemThreshold ?? 0) - (b.groupItemThreshold ?? 0),
-        )
-      : null;
-
-  if (sortedDrawOutcomes) {
-    const [homeOutcome, drawOutcome, awayOutcome] = sortedDrawOutcomes;
-    const homeToken = getTeamToken(homeOutcome.tokens, game.homeTeam);
-    const drawToken = getDrawToken(drawOutcome.tokens) ?? drawOutcome.tokens[0];
-    const awayToken = getTeamToken(awayOutcome.tokens, game.awayTeam);
-
-    return {
-      ...fallbackResult,
-      home: homeToken ? { outcome: homeOutcome, token: homeToken } : undefined,
-      draw: drawToken ? { outcome: drawOutcome, token: drawToken } : undefined,
-      away: awayToken ? { outcome: awayOutcome, token: awayToken } : undefined,
-    };
-  }
-
-  if (isTeamToAdvance && primaryOutcomes.length >= 2) {
-    const homeOutcome = getTeamOutcome(primaryOutcomes, game.homeTeam, 0);
-    const awayOutcome = getTeamOutcome(
-      primaryOutcomes,
-      game.awayTeam,
-      1,
-      homeOutcome,
-    );
-    const homeToken = homeOutcome
-      ? getTeamToken(homeOutcome.tokens, game.homeTeam)
-      : undefined;
-    const awayToken = awayOutcome
-      ? getTeamToken(
-          awayOutcome.tokens,
-          game.awayTeam,
-          [homeToken?.id].filter((id): id is string => Boolean(id)),
-        )
-      : undefined;
-
-    return {
-      ...fallbackResult,
-      home:
-        homeOutcome && homeToken
-          ? { outcome: homeOutcome, token: homeToken }
-          : undefined,
-      away:
-        awayOutcome && awayToken
-          ? { outcome: awayOutcome, token: awayToken }
-          : undefined,
-    };
-  }
-
-  const homeToken = getTeamToken(firstOutcome.tokens, game.homeTeam);
-  const drawToken =
-    showDraw && !isTeamToAdvance
-      ? getDrawToken(firstOutcome.tokens)
-      : undefined;
-  const awayToken = getTeamToken(
-    firstOutcome.tokens,
-    game.awayTeam,
-    [homeToken?.id, drawToken?.id].filter((id): id is string => Boolean(id)),
-  );
-
-  return {
-    ...fallbackResult,
-    home: homeToken ? { outcome: firstOutcome, token: homeToken } : undefined,
-    draw: drawToken ? { outcome: firstOutcome, token: drawToken } : undefined,
-    away: awayToken ? { outcome: firstOutcome, token: awayToken } : undefined,
-  };
 };
