+++ app/components/UI/Predict/providers/polymarket/PolymarketProvider.ts
@@ -18,10 +18,8 @@ import {
   isSmartContractAddress,
 } from '../../../../../util/transactions';
 import { PREDICT_CONSTANTS, PREDICT_ERROR_CODES } from '../../constants/errors';
-import {
-  filterSupportedLeagues,
-  getPrimarySportsCardOutcomes,
-} from '../../constants/sports';
+import { filterSupportedLeagues } from '../../constants/sports';
+import { getPrimarySportsCardOutcomes , resolveWorldCupFeedEvents } from './sportsUtils';
 import { PREDICT_ACTIVITY_PAGE_SIZE } from '../../constants/transactions';
 import { SERIES_MAX_EVENTS } from '../../utils/series';
 import {
@@ -168,7 +166,6 @@ import {
   waitForDepositWalletDeployed,
   waitForDepositWalletTransaction,
 } from './depositWallet';
-import { resolveWorldCupFeedEvents } from './sportsUtils';
 
 export type SignTypedMessageFn = (
   params: TypedMessageParams,
